# api/docs/swagger.yaml
definitions:
  controllers.ContainerStatusResponse:
    properties:
      container_id:
        type: string
      created:
        type: string
      is_running:
        type: boolean
      last_used:
        type: string
      port:
        type: integer
      status:
        type: string
      user_id:
        type: string
    type: object
  controllers.CreateInvoiceRequest:
    properties:
      amount:
        minimum: 0
        type: number
      month:
        maximum: 12
        minimum: 1
        type: integer
      transaction_hash:
        type: string
      user_id:
        type: string
      year:
        minimum: 2020
        type: integer
    required:
    - amount
    - month
    - transaction_hash
    - user_id
    - year
    type: object
  controllers.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
      timestamp:
        type: string
    type: object
  controllers.GenerateMonthlyInvoiceRequest:
    properties:
      amount:
        minimum: 0
        type: number
      month:
        maximum: 12
        minimum: 1
        type: integer
      payment_validation_url:
        type: string
      year:
        minimum: 2020
        type: integer
    required:
    - amount
    - month
    - payment_validation_url
    - year
    type: object
  controllers.InvoiceResponse:
    properties:
      amount:
        type: number
      created_at:
        type: string
      id:
        type: string
      is_paid:
        type: boolean
      last_active:
        type: string
      month:
        type: integer
      paid_at:
        type: string
      transaction_hash:
        type: string
      user_id:
        type: string
      year:
        type: integer
    type: object
  controllers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  controllers.LoginResponse:
    properties:
      container_id:
        type: string
      email:
        type: string
      last_active:
        type: string
      token:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
  controllers.MessageRequest:
    properties:
      message:
        type: string
    required:
    - message
    type: object
  controllers.MessageResponse:
    properties:
      container_id:
        type: string
      processed_at:
        type: string
      response:
        type: string
      user_id:
        type: string
    type: object
  controllers.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
      wallet_address:
        type: string
    required:
    - email
    - password
    - username
    type: object
  controllers.RegisterResponse:
    properties:
      container_id:
        type: string
      created_at:
        type: string
      email:
        type: string
      token:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
  controllers.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
  controllers.UpdateInvoiceRequest:
    properties:
      amount:
        minimum: 0
        type: number
      payment_validation_url:
        type: string
    type: object
  controllers.UpdateUserRequest:
    properties:
      email:
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    type: object
  controllers.UserProfileResponse:
    properties:
      container_id:
        type: string
      created_at:
        type: string
      email:
        type: string
      is_active:
        type: boolean
      last_active:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/containers/{containerID}:
    get:
      description: Returns container information by ID (only accessible by the container
        owner)
      parameters:
      - description: Container ID
        in: path
        name: containerID
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Container information
          schema:
            $ref: '#/definitions/controllers.ContainerStatusResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Container not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Get container by ID
      tags:
      - Container Management
  /api/v1/containers/message:
    post:
      consumes:
      - application/json
      description: Sends a message to the user's Jarvis agent running in their dedicated
        container
      parameters:
      - description: Message data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.MessageRequest'
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Message processed successfully
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Container not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Send message to user's agent
      tags:
      - Agent Communication
  /api/v1/containers/start:
    post:
      description: Starts the user's agent container if it's stopped
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Container started successfully
          schema:
            $ref: '#/definitions/controllers.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Container not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Failed to start container
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Start user's container
      tags:
      - Container Management
  /api/v1/containers/status:
    get:
      description: Returns the status of the user's agent container
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Container status
          schema:
            $ref: '#/definitions/controllers.ContainerStatusResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Container not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Get container status
      tags:
      - Container Management
  /api/v1/containers/stop:
    post:
      description: Stops the user's agent container
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Container stopped successfully
          schema:
            $ref: '#/definitions/controllers.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Container not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Failed to stop container
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Stop user's container
      tags:
      - Container Management
  /api/v1/invoices:
    get:
      description: Returns all invoices for the authenticated user
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of user invoices
          schema:
            items:
              $ref: '#/definitions/controllers.InvoiceResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Get user's invoices
      tags:
      - Invoice Management
    post:
      consumes:
      - application/json
      description: Creates a new invoice for a user
      parameters:
      - description: Invoice creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateInvoiceRequest'
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Invoice created successfully
          schema:
            $ref: '#/definitions/controllers.InvoiceResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "409":
          description: Invoice already exists
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Create a new invoice
      tags:
      - Invoice Management
  /api/v1/invoices/{invoiceID}:
    delete:
      description: Deletes an existing invoice (only accessible by invoice owner)
      parameters:
      - description: Invoice ID
        in: path
        name: invoiceID
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Invoice deleted successfully
          schema:
            $ref: '#/definitions/controllers.SuccessResponse'
        "400":
          description: Invalid invoice ID
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Invoice not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Delete invoice
      tags:
      - Invoice Management
    get:
      description: Returns a specific invoice by ID (only accessible by invoice owner)
      parameters:
      - description: Invoice ID
        in: path
        name: invoiceID
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Invoice details
          schema:
            $ref: '#/definitions/controllers.InvoiceResponse'
        "400":
          description: Invalid invoice ID
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Invoice not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Get invoice by ID
      tags:
      - Invoice Management
    put:
      consumes:
      - application/json
      description: Updates an existing invoice (only accessible by invoice owner)
      parameters:
      - description: Invoice ID
        in: path
        name: invoiceID
        required: true
        type: string
      - description: Updated invoice data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdateInvoiceRequest'
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Invoice updated successfully
          schema:
            $ref: '#/definitions/controllers.InvoiceResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Invoice not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Update invoice
      tags:
      - Invoice Management
  /api/v1/invoices/{invoiceID}/check-payment:
    post:
      description: Checks the payment validation URL and updates invoice status if
        paid
      parameters:
      - description: Invoice ID
        in: path
        name: invoiceID
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Payment status checked and updated
          schema:
            $ref: '#/definitions/controllers.InvoiceResponse'
        "400":
          description: Invalid invoice ID
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Invoice not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Check payment status
      tags:
      - Invoice Management
  /api/v1/invoices/all/paid:
    get:
      description: Returns all paid invoices in the system (admin only)
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of paid invoices
          schema:
            items:
              $ref: '#/definitions/controllers.InvoiceResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Get all paid invoices
      tags:
      - Invoice Management
  /api/v1/invoices/all/unpaid:
    get:
      description: Returns all unpaid invoices in the system (admin only)
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of unpaid invoices
          schema:
            items:
              $ref: '#/definitions/controllers.InvoiceResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Get all unpaid invoices
      tags:
      - Invoice Management
  /api/v1/invoices/generate:
    post:
      consumes:
      - application/json
      description: Creates a monthly invoice for the authenticated user
      parameters:
      - description: Monthly invoice data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.GenerateMonthlyInvoiceRequest'
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Monthly invoice created successfully
          schema:
            $ref: '#/definitions/controllers.InvoiceResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "409":
          description: Invoice already exists
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Generate monthly invoice
      tags:
      - Invoice Management
  /api/v1/invoices/stats:
    get:
      description: Returns statistics about invoices (admin only)
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Invoice statistics
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Get invoice statistics
      tags:
      - Invoice Management
  /api/v1/invoices/unpaid:
    get:
      description: Returns all unpaid invoices for the authenticated user
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of unpaid invoices
          schema:
            items:
              $ref: '#/definitions/controllers.InvoiceResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Get user's unpaid invoices
      tags:
      - Invoice Management
  /api/v1/invoices/user/{userID}:
    get:
      description: Returns all invoices for a specific user (admin only)
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of user invoices
          schema:
            items:
              $ref: '#/definitions/controllers.InvoiceResponse'
            type: array
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Get invoices by user ID
      tags:
      - Invoice Management
  /api/v1/users/{userID}:
    delete:
      description: Deletes the authenticated user's account and associated container
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
          schema:
            $ref: '#/definitions/controllers.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Delete user account
      tags:
      - User Management
    get:
      description: Returns user information by ID (only accessible by the user themselves)
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User information
          schema:
            $ref: '#/definitions/controllers.UserProfileResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Get user by ID
      tags:
      - User Management
    put:
      consumes:
      - application/json
      description: Updates the authenticated user's profile information
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: Updated user data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdateUserRequest'
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Updated user information
          schema:
            $ref: '#/definitions/controllers.UserProfileResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Update user information
      tags:
      - User Management
  /api/v1/users/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user with email and password, returns a token
      parameters:
      - description: User login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/controllers.LoginResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Login user
      tags:
      - Authentication
  /api/v1/users/profile:
    get:
      description: Returns the authenticated user's profile information
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User profile
          schema:
            $ref: '#/definitions/controllers.UserProfileResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Get user profile
      tags:
      - User Management
  /api/v1/users/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account and spins up a dedicated Docker container
        with a Jarvis agent instance
      parameters:
      - description: User registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            $ref: '#/definitions/controllers.RegisterResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Register a new user
      tags:
      - User Management
swagger: "2.0"